requirements.txt
google-cloud-firestore==0.29.0